refactor code into a util/audio module
- removes code not directly related to validation from
config/validators
- moves files_from_dir from io.audio into a util module, so that
  just the main (and relatively lengthy) 'to_spect' function
  lives in io.audio
- wraps evfuncs.load_cbin in a decorator that switches the returned
  argument order, so it's the same order as wavfile.read, i.e.
  sampling frequency is element 0 of the return tuple, and
  the actual data is element 1 of the return tuple
  + this was not a problem before with the 'built-in' evfuncs
    module that I tore out
- config/validators now just uses AUDIO_FORMAT_FUNC_MAP from util.audio
