+++ src/vak/util/__init__.py
@@ -1,3 +1,4 @@
+from . import audio
 from . import curvefit
 from . import data
 from . import dataset
